renamed auth feature folders
