+++ lib/src/features/authentication/data/fake_app_user.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:irrigazione_iot/src/features/authentication/domain/app_user.dart';
+import 'package:irrigazione_iot/src/features/authentication/model/app_user.dart';
 
 /// Fake user class used to simulate a user account on the backend
 /// * This class is implementation-specific and should only be used by the
