+++ lib/src/features/authentication/model/app_user.dart
